Propertize syntax with scheme-mode table for Scheme ranges
Apply default syntax-table to everything else. Right now, this just uses the
static `scheme-mode-syntax-table`. Calling `scheme-propertize-syntax` even
inside `with-syntax-table scheme-mode-syntax-table` doesn't seem to work, or at
least only works consistently.

This allows for any tooling that relies on the syntax table to work properly in
Scheme blocks, as long as it isn't confused by surrounding Lilypond code.

Also added a hook to turn on Geiser mode for Lilypond buffers, but only if the
`geiser-lilypond-guile` implementation is available — currently non-public.