REPL: major refactor
* Integrate REPL config with version selection and argument sets
* Don't enable full `geiser-mode` in LilyPond buffers. As of this commit, that
removes some functionality, notably autodoc, but Geiser trying to parse LilyPond
code using Scheme regexes likely leads to a variety of bugs. A better approach
is to expose only compatible Geiser functionality via lilypond-ts-mode features.
* Bugfix: various errors due to LilyPond mode setup polluting REPL buffer.
* Remove keywords from lilypond-ts-base to allow eval to live in REPL file
* Rename feature as integral part of `lilypond-ts` and remove conditional
checks. Apparently Geiser breaks if an implementation doesn't provide exactly
`geiser-<impl>`, so for now the file provides both feature names. Feels wrong
to have a package component with a different prefix, or to use a file name with
a different prefix than the contents, since it really isn't separate from the
rest of lilypond-ts-mode. But this is also ugly. Unclear the best thing to do.