Nav: rough initial rewrite to use overlays for lookups
The old approach of using a lookup table for navigation is brittle, because the
locations in the table become incorrect quickly during editing. It is awkward to
require recompiling in the middle of a vertical edit, since such edits may leave
the score in a rhythmically inconsistent state.

The new approach uses a lookup table only for traversing between input files,
and instead directly searches across overlays. In order to do this more
efficiently, separate overlays are created to span sequential segments (logic
that was previously on the Scheme side).

Uncommented tests pass for this commit.