Scheme mode: major update to handle refactored grammar
In order to handle embedded LilyPond, the Scheme grammar now exactly mirrors
the LilyPond grammar in that the embedded grammar is fully built-in.

Hence ranges are NO LONGER NEEDED, and LilyPond/Scheme modes can have identical
code, with just a flag to flip which language is at the top level.

Unwinding the complexity of the existing ranges-based code means this is a large
commit.

Further simplification is likely possible in consolidating more into the common
parent mode. Syntax propertize can also almost certainly be simplified and made
more efficient.

The LilyPond mode is mostly functionally unaffected, with the exception that the
grammar repo location has changed.

There is one font lock rule that only works in LilyPond mode, due to a grammar
quirk. Likely this is a grammar bug that will be fixed.